+++ lib/android_location.dart
@@ -4,4 +4,8 @@
 /// related services.
 ///
 /// See: https://developer.android.com/reference/android/location/package-summary
+///
+/// {@canonicalFor location.Location}
 library android_location;
+
+export 'src/location/location.dart' show Location;