+++ lib/src/location/location.dart
@@ -1,0 +1,9 @@
+/* This is free and unencumbered software released into the public domain. */
+
+/// A data class representing a geographic location.
+///
+/// A location can consist of a latitude, longitude, timestamp, and other
+/// information such as bearing, altitude and velocity.
+///
+/// See: https://developer.android.com/reference/android/location/Location
+abstract class Location {}