+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.3.3] - 2019-02-15
 ### Added
+- `android_location.Location` class stub
 
 ## [0.3.2] - 2019-02-15
 ### Added